+++ packages/vendure-plugin-accept-blue/src/service/accept-blue-service.ts
@@ -2,6 +2,7 @@
 import { Inject, Injectable, OnApplicationBootstrap } from '@nestjs/common';
 import { ModuleRef } from '@nestjs/core';
 import {
+  assertFound,
   Customer,
   CustomerService,
   EntityHydrator,
@@ -44,6 +45,7 @@ import {
   NoncePaymentMethodInput,
   SavedPaymentMethodInput,
   StorefrontKeys,
+  AcceptBlueCardPaymentMethod,
 } from '../types';
 import {
   getNrOfBillingCyclesLeft,
@@ -56,13 +58,16 @@ import {
 import { AcceptBlueClient } from './accept-blue-client';
 import { acceptBluePaymentHandler } from './accept-blue-handler';
 import {
+  AcceptBlueCheckPaymentMethod,
   AcceptBluePaymentMethodQuote,
   AcceptBlueRefundResult,
   AcceptBlueSubscription,
   AcceptBlueSurcharges,
   AcceptBlueTransaction,
+  UpdateAcceptBlueCardPaymentMethodInput,
+  UpdateAcceptBlueCheckPaymentMethodInput,
   UpdateAcceptBlueSubscriptionInput,
-} from './generated/graphql';
+} from '../api/generated/graphql';
 
 @Injectable()
 export class AcceptBlueService implements OnApplicationBootstrap {
@@ -201,10 +206,12 @@ export class AcceptBlueService implements OnApplicationBootstrap {
     let paymentMethod: AcceptBluePaymentMethod;
     if (isSavedPaymentMethod(input as SavedPaymentMethodInput)) {
       const foundMethod = await client.getPaymentMethod(
-        acceptBlueCustomerId,
         (input as SavedPaymentMethodInput).paymentMethodId
       );
-      if (!foundMethod) {
+      if (
+        !foundMethod ||
+        foundMethod.customer_id != String(acceptBlueCustomerId)
+      ) {
         throw new UserInputError(
           `No Accept Blue payment method found with id ${
             (input as SavedPaymentMethodInput).paymentMethodId
@@ -412,6 +419,35 @@ export class AcceptBlueService implements OnApplicationBootstrap {
     return recurringSchedules;
   }
 
+  async updateCardPaymentMethod(
+    ctx: RequestContext,
+    input: UpdateAcceptBlueCardPaymentMethodInput
+  ): Promise<AcceptBlueCardPaymentMethod> {
+    const client = await this.getClientForChannel(ctx);
+    await this.isUpdatePaymentMethodAllowed(ctx, client, input.id, 'card');
+    return (await client.updatePaymentMethod(input.id, {
+      name: input.name || undefined,
+      expiry_month: input.expiry_month ?? undefined,
+      expiry_year: input.expiry_year ?? undefined,
+      avs_address: input.avs_address ?? undefined,
+      avs_zip: input.avs_zip ?? undefined,
+    })) as AcceptBlueCardPaymentMethod;
+  }
+
+  async updateCheckPaymentMethod(
+    ctx: RequestContext,
+    input: UpdateAcceptBlueCheckPaymentMethodInput
+  ): Promise<AcceptBlueCheckPaymentMethod> {
+    const client = await this.getClientForChannel(ctx);
+    await this.isUpdatePaymentMethodAllowed(ctx, client, input.id, 'check');
+    return (await client.updatePaymentMethod(input.id, {
+      name: input.name ?? undefined,
+      routing_number: input.routing_number ?? undefined,
+      account_type: input.account_type ?? undefined,
+      sec_code: input.sec_code ?? undefined,
+    })) as AcceptBlueCheckPaymentMethod;
+  }
+
   async updateSubscription(
     ctx: RequestContext,
     input: UpdateAcceptBlueSubscriptionInput
@@ -451,6 +487,53 @@ export class AcceptBlueService implements OnApplicationBootstrap {
       new AcceptBlueSubscriptionEvent(ctx, subscription, 'updated', input)
     );
     return subscription;
+  }
+
+  /**
+   * Check if the logged in user is allowed to update the payment method.
+   * Throws an error if not allowed
+   *
+   * 1. Is admin > all good
+   * 2. Is shop API > requires logged in customer and ownership of the payment method
+   */
+  private async isUpdatePaymentMethodAllowed(
+    ctx: RequestContext,
+    client: AcceptBlueClient,
+    paymentMethodId: number,
+    shouldBe: 'card' | 'check'
+  ): Promise<void> {
+    const existingMethod = await client.getPaymentMethod(paymentMethodId);
+    if (!existingMethod) {
+      throw new UserInputError(
+        `No Accept Blue payment method found with id ${paymentMethodId}`
+      );
+    }
+    if (existingMethod.payment_method_type !== shouldBe) {
+      throw new UserInputError(
+        `Payment method '${paymentMethodId}' is not a ${shouldBe} payment method`
+      );
+    }
+    // if Shop API, we check if the logged in user is owner of the payment method
+    if (ctx.apiType === 'shop') {
+      if (!ctx.activeUserId) {
+        throw new ForbiddenError();
+      }
+      const customer = await assertFound(
+        this.customerService.findOneByUserId(ctx, ctx.activeUserId)
+      );
+      if (!customer.customFields.acceptBlueCustomerId) {
+        throw new UserInputError(
+          `Customer '${customer.emailAddress}' (${customer.id}) is not linked to an Accept Blue customer`
+        );
+      }
+      if (
+        existingMethod.customer_id !=
+        String(customer.customFields.acceptBlueCustomerId)
+      ) {
+        // Given ID is not a payment method of the customer
+        throw new ForbiddenError();
+      }
+    }
   }
 
   /**
@@ -677,7 +760,6 @@ export class AcceptBlueService implements OnApplicationBootstrap {
     }
     return acceptBlueMethod[0];
   }
-
   /**
    * Handle incoming webhooks from Accept Blue
    */